+++ sources/plug_bullet/src/luna/bind_btGeneric6DofConstraint.cpp
@@ -1454,7 +1454,7 @@ public:
 		}
 		btVector3 & ax1=*ax1_ptr;
 		int rotational=(int)lua_tointeger(L,12);
-		int rotAllowed=luatop>12 ? (int)lua_tointeger(L,13) : false;
+		int rotAllowed=luatop>12 ? (int)lua_tointeger(L,13) : (int)false;
 
 		btGeneric6DofConstraint* self=Luna< btTypedObject >::checkSubType< btGeneric6DofConstraint >(L,1);
 		if(!self) {
@@ -1516,7 +1516,7 @@ public:
 
 		int num=(int)lua_tointeger(L,2);
 		float value=(float)lua_tonumber(L,3);
-		int axis=luatop>3 ? (int)lua_tointeger(L,4) : -1;
+		int axis=luatop>3 ? (int)lua_tointeger(L,4) : (int)-1;
 
 		btGeneric6DofConstraint* self=Luna< btTypedObject >::checkSubType< btGeneric6DofConstraint >(L,1);
 		if(!self) {
@@ -1538,7 +1538,7 @@ public:
 		int luatop = lua_gettop(L);
 
 		int num=(int)lua_tointeger(L,2);
-		int axis=luatop>2 ? (int)lua_tointeger(L,3) : -1;
+		int axis=luatop>2 ? (int)lua_tointeger(L,3) : (int)-1;
 
 		btGeneric6DofConstraint* self=Luna< btTypedObject >::checkSubType< btGeneric6DofConstraint >(L,1);
 		if(!self) {
@@ -1776,7 +1776,7 @@ public:
 
 		int num=(int)lua_tointeger(L,2);
 		float value=(float)lua_tonumber(L,3);
-		int axis=luatop>3 ? (int)lua_tointeger(L,4) : -1;
+		int axis=luatop>3 ? (int)lua_tointeger(L,4) : (int)-1;
 
 		btGeneric6DofConstraint* self=Luna< btTypedObject >::checkSubType< btGeneric6DofConstraint >(L,1);
 		if(!self) {
@@ -1798,7 +1798,7 @@ public:
 		int luatop = lua_gettop(L);
 
 		int num=(int)lua_tointeger(L,2);
-		int axis=luatop>2 ? (int)lua_tointeger(L,3) : -1;
+		int axis=luatop>2 ? (int)lua_tointeger(L,3) : (int)-1;
 
 		btGeneric6DofConstraint* self=Luna< btTypedObject >::checkSubType< btGeneric6DofConstraint >(L,1);
 		if(!self) {
