+++ sources/plug_bullet/src/luna/bind_btPoint2PointConstraint.cpp
@@ -691,7 +691,7 @@ public:
 
 		int num=(int)lua_tointeger(L,2);
 		float value=(float)lua_tonumber(L,3);
-		int axis=luatop>3 ? (int)lua_tointeger(L,4) : -1;
+		int axis=luatop>3 ? (int)lua_tointeger(L,4) : (int)-1;
 
 		btPoint2PointConstraint* self=Luna< btTypedObject >::checkSubType< btPoint2PointConstraint >(L,1);
 		if(!self) {
@@ -713,7 +713,7 @@ public:
 		int luatop = lua_gettop(L);
 
 		int num=(int)lua_tointeger(L,2);
-		int axis=luatop>2 ? (int)lua_tointeger(L,3) : -1;
+		int axis=luatop>2 ? (int)lua_tointeger(L,3) : (int)-1;
 
 		btPoint2PointConstraint* self=Luna< btTypedObject >::checkSubType< btPoint2PointConstraint >(L,1);
 		if(!self) {
@@ -949,7 +949,7 @@ public:
 
 		int num=(int)lua_tointeger(L,2);
 		float value=(float)lua_tonumber(L,3);
-		int axis=luatop>3 ? (int)lua_tointeger(L,4) : -1;
+		int axis=luatop>3 ? (int)lua_tointeger(L,4) : (int)-1;
 
 		btPoint2PointConstraint* self=Luna< btTypedObject >::checkSubType< btPoint2PointConstraint >(L,1);
 		if(!self) {
@@ -971,7 +971,7 @@ public:
 		int luatop = lua_gettop(L);
 
 		int num=(int)lua_tointeger(L,2);
-		int axis=luatop>2 ? (int)lua_tointeger(L,3) : -1;
+		int axis=luatop>2 ? (int)lua_tointeger(L,3) : (int)-1;
 
 		btPoint2PointConstraint* self=Luna< btTypedObject >::checkSubType< btPoint2PointConstraint >(L,1);
 		if(!self) {
