+++ sources/plug_bullet/src/luna/bind_btPolyhedralConvexAabbCachingShape.cpp
@@ -723,7 +723,7 @@ public:
 
 		int luatop = lua_gettop(L);
 
-		int shiftVerticesByMargin=luatop>1 ? (int)lua_tointeger(L,2) : 0;
+		int shiftVerticesByMargin=luatop>1 ? (int)lua_tointeger(L,2) : (int)0;
 
 		btPolyhedralConvexAabbCachingShape* self=Luna< btCollisionShape >::checkSubType< btPolyhedralConvexAabbCachingShape >(L,1);
 		if(!self) {
