+++ sources/plug_bullet/src/luna/bind_btPolyhedralConvexShape.cpp
@@ -377,7 +377,7 @@ public:
 
 		int luatop = lua_gettop(L);
 
-		int shiftVerticesByMargin=luatop>1 ? (int)lua_tointeger(L,2) : 0;
+		int shiftVerticesByMargin=luatop>1 ? (int)lua_tointeger(L,2) : (int)0;
 
 		btPolyhedralConvexShape* self=Luna< btCollisionShape >::checkSubType< btPolyhedralConvexShape >(L,1);
 		if(!self) {
@@ -1008,7 +1008,7 @@ public:
 
 		int luatop = lua_gettop(L);
 
-		int shiftVerticesByMargin=luatop>1 ? (int)lua_tointeger(L,2) : 0;
+		int shiftVerticesByMargin=luatop>1 ? (int)lua_tointeger(L,2) : (int)0;
 
 		btPolyhedralConvexShape* self=Luna< btCollisionShape >::checkSubType< btPolyhedralConvexShape >(L,1);
 		if(!self) {
