+++ sources/plug_bullet/src/luna/bind_btSliderConstraint.cpp
@@ -2589,7 +2589,7 @@ public:
 
 		int num=(int)lua_tointeger(L,2);
 		float value=(float)lua_tonumber(L,3);
-		int axis=luatop>3 ? (int)lua_tointeger(L,4) : -1;
+		int axis=luatop>3 ? (int)lua_tointeger(L,4) : (int)-1;
 
 		btSliderConstraint* self=Luna< btTypedObject >::checkSubType< btSliderConstraint >(L,1);
 		if(!self) {
@@ -2611,7 +2611,7 @@ public:
 		int luatop = lua_gettop(L);
 
 		int num=(int)lua_tointeger(L,2);
-		int axis=luatop>2 ? (int)lua_tointeger(L,3) : -1;
+		int axis=luatop>2 ? (int)lua_tointeger(L,3) : (int)-1;
 
 		btSliderConstraint* self=Luna< btTypedObject >::checkSubType< btSliderConstraint >(L,1);
 		if(!self) {
@@ -2765,7 +2765,7 @@ public:
 
 		int num=(int)lua_tointeger(L,2);
 		float value=(float)lua_tonumber(L,3);
-		int axis=luatop>3 ? (int)lua_tointeger(L,4) : -1;
+		int axis=luatop>3 ? (int)lua_tointeger(L,4) : (int)-1;
 
 		btSliderConstraint* self=Luna< btTypedObject >::checkSubType< btSliderConstraint >(L,1);
 		if(!self) {
@@ -2787,7 +2787,7 @@ public:
 		int luatop = lua_gettop(L);
 
 		int num=(int)lua_tointeger(L,2);
-		int axis=luatop>2 ? (int)lua_tointeger(L,3) : -1;
+		int axis=luatop>2 ? (int)lua_tointeger(L,3) : (int)-1;
 
 		btSliderConstraint* self=Luna< btTypedObject >::checkSubType< btSliderConstraint >(L,1);
 		if(!self) {
