+++ sources/plug_bullet/src/luna/bind_btTetrahedronShapeEx.cpp
@@ -722,7 +722,7 @@ public:
 
 		int luatop = lua_gettop(L);
 
-		int shiftVerticesByMargin=luatop>1 ? (int)lua_tointeger(L,2) : 0;
+		int shiftVerticesByMargin=luatop>1 ? (int)lua_tointeger(L,2) : (int)0;
 
 		btTetrahedronShapeEx* self=Luna< btCollisionShape >::checkSubType< btTetrahedronShapeEx >(L,1);
 		if(!self) {
